Added todo Authorization

+++ backend/routes/todos.js
@@ -4,17 +4,16 @@ const router = express.Router();
 const checkAuth = require('../middleware/checkAuth');
 
 router.post('/create-todo', checkAuth, (req, res) => {
-    
-    const todo = new Todo({todo: req.body.todo, checked: false});
-    console.log(req.headers.authorization)
+    const todo = new Todo({ todo: req.body.todo, checked: false, creator: req.decodedToken.id });
 
     todo.save()
     .then(document => {
-
         res.status(201).json({
             message: 'Successfully posted the todo!',
             todo: document
+
         })
+
     })
     .catch(err => {
 
@@ -23,13 +22,12 @@ router.post('/create-todo', checkAuth, (req, res) => {
 })
 
 router.get('/get-todos', checkAuth, (req, res) => {
-
-    Todo.find()
-    .then( documents => {
-
+    Todo.find({ creator: req.decodedToken.id })
+    .then(documents => {
             res.status(200).json({
                 message: 'successfullt fetched todos!',
                 todos: documents
+
             })
         }
     )
@@ -40,15 +38,15 @@ router.get('/get-todos', checkAuth, (req, res) => {
 })
 
 router.put('/todo-check', checkAuth, (req, res) => {
-
-    Todo.updateOne({_id: req.query.id}, {checked: req.body.checked, _id: req.body.todoItem.id})
+    Todo.updateOne(
+        { _id: req.query.id, creator: req.decodedToken.id }, 
+        {checked: req.body.checked, _id: req.body.todoItem.id}
+        )
     .then(document => {
-
-        console.log('todo: ', document)
-        
         res.status(200).json({
             message: 'Successfully edited the todo!',
             todo: document
+
         });
     })
     .catch(err => {
@@ -60,15 +58,19 @@ router.put('/todo-check', checkAuth, (req, res) => {
 })
 
 router.delete('/todo-delete', checkAuth, (req , res) => {
-    console.log('headers: ', req.headers.authorization)
-
-    Todo.deleteOne({ _id: req.query.id })
+    Todo.deleteOne({ _id: req.query.id, creator: req.decodedToken.id })
     .then(result => {
-        
-        res.status(202).json({
-            message: 'Successfully deleted todo!'
+        if(result.deletedCount) {
+            res.status(202).json({
+                message: 'Successfully deleted todo!'
+    
+            }) 
 
-        }) 
+        } else {
+            res.status(401).json({ message: 'Not Authorized!' });
+
+        }
+
     })
